Stage (#177)
* In this iteration we added the frontend. It consists of one page of static html. (#173)

This page uses this new importmap type that allows to define the versions of vue and vuetify that we need, which are loaded in the script section.
So we end up with a single page of HTML that has the html, the javascript and the styles required to run the whole website!
In this iteration we also changed the lambda function from a POST to a GET.

Signed-off-by: Daniel Mermelstein <danmermel@gmail.com>

* In this iteration we are adding a GH workflow to deploy the production site to GH pages (#174)

Will it work?

Signed-off-by: Daniel Mermelstein <danmermel@gmail.com>

* Deploy prod (#176)

* In this iteration we are adding a GH workflow to deploy the production site to GH pages
Will it work?

Signed-off-by: Daniel Mermelstein <danmermel@gmail.com>

* In this iteration we fixed the thing that was causing the CORS error.
Here's a good explainer of what CORS is all about https://www.reddit.com/r/learnjavascript/comments/pzuy6h/can_someone_explain_to_me_what_is_crossorigin/?rdt=57080

Signed-off-by: Daniel Mermelstein <danmermel@gmail.com>

---------

Signed-off-by: Daniel Mermelstein <danmermel@gmail.com>

---------

Signed-off-by: Daniel Mermelstein <danmermel@gmail.com>